+++ docs/src/config/navigation.ts
@@ -35,6 +35,7 @@ export const navigationConfig: Navigation = [
       { label: 'DXDashboard', url: '/components/extended/DXDashboard' },
       { label: 'DXDashboardNavbar', url: '/components/extended/DXDashboardNavbar' },
       { label: 'DXDashboardSidebar', url: '/components/extended/DXDashboardSidebar' },
+      { label: 'DXCurrencyInput', url: '/components/extended/DXCurrencyInput' },
       { label: 'DXForm', url: '/components/extended/DXForm' },
       { label: 'DXField', url: '/components/extended/DXField' },
       { label: 'DXRepeater', url: '/components/extended/DXRepeater' },